refactor(CoreData): moved files to CoreData folder
