+++ CoreData/Customer+CoreDataProperties.swift
