+++ CoreData/Service+CoreDataProperties.swift
